+++ apps/client/src/environments/extracts-hash.ts
@@ -1,1 +1,1 @@
-export const extractsHash = `f323f0aa3f72a300765f47dce920b01b8df06b3f`;
+export const extractsHash = `33d78f7dc72e96f266e1c092d1d69b74e08d9e33`;