+++ apps/client/src/environments/patch-notes.ts
@@ -1,9 +1,20 @@
 export const patchNotes = `### Bug Fixes
 
-* **db:** fixed missing trade sources from 6.1 update.
-* **profile:** removed fishing rankings for now and fixed missing job icons.
+* **allagan-reports:** better UX for spearfishing input.
+* **community-rotations:** add 50 durability as possible value.
+* **db:** add missing voyages from 6.1 update.
+* **desktop:** fixed submarine progression capture.
+* **desktop:** possible fix for list overlay freezing entirely.
+* **search:** fixed gathering search for korean region.
+* **simulator:** fixed name change not applied to display.
 
 
 ### Features
 
-* **community-lists:** now sorting by most recent app version and then by # of forks.`;
+* **alarms:** added hookset icon to sidebar and overlay.
+* **core:** added server name next ot character name in quick switch.
+* **db:** 500+ allagan reports imported.
+* **db:** new allagan reports import.
+* **lists:** new button to select items that are not completed in list split popup.
+* **profile:** new characters management box in profile editor.
+* **simulator:** added icons next to selected food and medicine.`;